Compile in two steps, tsc + webpack
Instead of compiling the typescript and webpacking in one go, do the
process in two steps:

- Compile ts files to corresponding js files
- Webpack the js files

I find it useful to see the intermediary .js files and be able to
inspect them.  Also, debugging the webpack bundle does not work
currently.  Debugging the intermediary js files work.  We'll of couse
want to be able to debug the bundle, but this gives us a way to do it in
the mean time.

A functional change of this patch is that the bundle is now generated
to dist/bundle.js, which is a kind of de facto standard place for it.
it can be renamed to anything when distributed.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>