Add logging with bunyan
The GDBDebugSession class currently inherits from LoggingDebugSession
(from vscode-debugadapter), but I am pretty sure logging is not
functional.  In particular, we pass it a relative file name, but the
logger code requires it to be absolute:

    https://github.com/Microsoft/vscode-debugadapter-node/blob/ac2d3a31cf5e71999e28c191f0b80a56439954bc/adapter/src/logger.ts#L144

Therefore I conclude it's not used at the moment.

Instead of using LoggingDebugSession, which is quite rudimentary, I
suggest using a dedicated logging library.  This will give us a bit more
control on what we log, where it goes, etc.  I have tried winston and
bunyan, two popular libraries, but got better and easier results with
the bunyan, so I went with that.

With this patch, there is a default stream to stderr which shows
warnings and above.  Passing `logFile` in a launch request enables a log
file with levels `debug` and above to be written, with all the DA and MI
communications.

Note that the logs are written as json files, and the `bunyan` CLI tool
can be used to display them nicely,

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>