Fix calls to sendErrorResponse
At GDBBackend.ts:78, we reject a Promise with an object.  That rejected
promise may become an exception, caught for example in
setBreakPointsRequest at GDBDebugSession.ts:178.  That object is then
passed to sendErrorResponse.  The problem is that sendErrorResponse
expects to receive a string.  So it fails with:

(node:2643) UnhandledPromiseRejectionWarning: TypeError: format.replace is not a function
    at Function.formatPII (/home/emaisin/src/cdt-gdb-adapter/node_modules/vscode-debugadapter/lib/debugSession.js:669:23)
    at GDBDebugSession.sendErrorResponse (/home/emaisin/src/cdt-gdb-adapter/node_modules/vscode-debugadapter/lib/debugSession.js:310:41)
    at GDBDebugSession.<anonymous> (/home/emaisin/src/cdt-gdb-adapter/out/GDBDebugSession.js:137:22)
    at Generator.throw (<anonymous>)
    at rejected (/home/emaisin/src/cdt-gdb-adapter/out/GDBDebugSession.js:5:65)
    at <anonymous>
    at process._tickCallback (internal/process/next_tick.js:189:7)

Unfortunately, exceptions in Typescript are not typed, so it's difficult to
make sure what kind of object we caught.  To circumvent this, I suggest
we make sure to always throw and reject with Error (or subclasses of
it) containing a message.  This way, we know that we can use err.message
as an error message.

With this fix, a failed request is properly handled, we send a response
with "success":"false" to the client.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>