Add 'Set Value' Functionality for Variables
Adds the ability to right-click > Set Value on most variables.

Limitations:
-Defined length arrays display the value of the array length when first
created, and the value of the first array entry that gets updated after
subsequent checks. It would be more useful to display the address the
array is located at in memory.
-Parent and child entries in the Variables tree are not updated when a
value is set, only the variable that was modified. This is purposefully
defined behavior by VS Code but not clear if this is modifiable.
-There exists a hook in the Debug Protocol to set expressions. I have
not
yet found a way to trigger a SetExpressionRequest.

Refs eclipse-cdt#15

Change-Id: I6d3c072bf9e6438cd2f5371f87e8cfeb53da6ca5
Signed-off-by: Jonathan Williams <jonwilliams@blackberry.com>