Use -file-exec-and-symbol to specify binary
We currently pass the program name on the GDB command line.  If, for
example, the program does not exist, the failure is silent (GDB outputs
an error as CLI output, but that's it).  The failure only manifests
itself when we try to run, but at this point the error is not precise:

  Error: No executable file specified.
  Use the "file" or "exec-file" command.

By using -file-exec-and-symbols, we are able to get a more precise
error:

  /does/not/exist: No such file or directory.

Fixes #29.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>