Add attach test
Add some simple "attach" tests, one when we successfully attach and one
when we provide a non-existent pid.

I kept it really simple, because I don't yet know what is the right
behavior we should test for.  For example, if the client sends a
configurationDone after a failed attach request, what should we reply?
So the tests can be improved at a later point.

I started to move things out of the existing test to a shared library of
test utilities.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>