Clean up termination.
On exit we need to interrupt gdb for the exit command to work
without hanging. For remote targets, we also need to kill the inferior.
That should be fine for local targets as well.

This change adds interrupt to the backend and then changes
the disconnnect request to follow the steps above.

Also fix up logger. When subclassing we had trouble getting the right
logger setup.