Send error message if evaluate is missing frameId
We currently don't support evaluating requests that lack a frameId
(normally, we should evaluate them in the global scope).

Instead of sending back that the request has succeeded with an error
message as the value, return an error.  Otherwise, the client doesn't
really have a way to notice the failure, since what it receives looks
like a regular value.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>