Change behavior of evaluate request handling
As discussed in #47, this patch changes how evaluate requests are
handled.  Currently, they are handled as GDB commands.  The intent of
the request is more to evaluate expressions in the target language.

This is already what we do when context is 'watch'.  This patch makes it
so we handle it the same way regardless of the context.

A simple test for the evaluate request is added.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>