Add support for arbritrary address expressions in memory read requests
Currently, it's only possible to pass a number for the address from
which we want to read memory, in our cdt-gdb-adapter/Memory request.  It
can be useful for the user to type expressions that resolve to an
address, such as "&myVariable" or "myPointer".  By accepting these
expressions (and forwarding them directly to GDB), we allow frontends to
pass directly what the user types in.

This indirectly fixes #66, where we wouldn't support full 64 bits
addresses (because node's integers are 53 bits).  We bypass it, since
all expressions would now be passed as strings, even literal addresses.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>