Some changes to evaluate request  (#69)
* Change behavior of evaluate request handling

As discussed in #47, this patch changes how evaluate requests are
handled.  Currently, they are handled as GDB commands.  The intent of
the request is more to evaluate expressions in the target language.

This is already what we do when context is 'watch'.  This patch makes it
so we handle it the same way regardless of the context.

A simple test for the evaluate request is added.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>

* Quote expressions passed to -var-create

Creating variable objects from expressions containing spaces (such as
"2 + 2") does not work currently.

Quote the expression passed to -var-create to handle this use case.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>

* Send error message if evaluate is missing frameId

We currently don't support evaluating requests that lack a frameId
(normally, we should evaluate them in the global scope).

Instead of sending back that the request has succeeded with an error
message as the value, return an error.  Otherwise, the client doesn't
really have a way to notice the failure, since what it receives looks
like a regular value.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>

* Send error when evaluate fails

When processing an evaluate request with an invalid expression (such as
"2 + "), we send back a "success" with an error message in the body.
This doesn't let the client handle the failure correctly.  Instead, send
back an error.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>