Test with and without `openGdbConsole` option
Because this features relies on the client opening GDB on behalf of the
debug adapter, and that there will be 2 uis running at the same time, we
want to ensure that everything works, first the standard way, and then
by delegating GDB's spawning to the client.

Problem is that the `DebugClient` from `vscode-debugprotocol` doesn't
seem to support reverse requests (a request from the debug adapter to
the client).

So I extended the `DebugClient` class to support the feature.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>