Stow away the logger to make sure subclasses use the same one.
In our plugin that extends the cdt adapter, we found that we were
unable to access the same logger that this extension was using.
It seemed to create a second logger. That made it hard to properly
set up the logging level. This stows a reference to it in the parent.