Get pause and setBreakpoint while running working.
Still not super clear this is the right approach but it works in
the small use case I tried. Pause request simply calls
kill('SIGINT")  on the child process. Works on Linux at least.

Also calls pause on the setBreakpoints request and waits for
the stopped async message. Then sets the breakpoints and
continues.