Enable mi-async
GDB doesn't listen to the MI channel by default when the debuggee is
running. For that to work, we need to enable `mi-async`. Since GDB only
answered when the debuggee stopped, most requests assuming that threads
were stopped used to succeed. But now that commands are evaluated more
quickly, I had to adapt the tests to wait for the program to stop before
doing the requests.

By default setting the `mi-async` option, fallback to setting
`target-async` on error.

Change the content of our `StoppedEvent` caused by signals:
- `reason` is always `pause`.
- `text` is the signal name.
- `description` is the meaning of the signal, if available.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>