Run make once before all tests run
This is done by moving invocation of make to a single
root-level mocha before hook. The timeout of this hook
is also increased to handle the relatively slow windows
build time.

This achieves two key fixes, one running make on windows
repeatedly sometimes fails with write access issues on .exe.
The other is that make is slow on windows, so running
repeatedly (and unnecessarily) is slow.