Gracefully handle breakpoints on unknown files and lines
Use Breakpoint.message instead of fully failing break insertion

In cases the breakpoint is rejected by GDB -break-insert is causing
an exception, that is being propagated up to a sendErrorResponse instead
of simply failing to verify that one breakpoint.

Note - if a breakpoint fails to insert because the Source is bad, instead
of the line number being bad some optimization could be made to set
verified to false on all the breakpoints.