pty: change the way we interact with ptys
The hack using sockets was causing troubles where some lines were not
correctly sent. This commit tries to simplify the pty fd handling.

There is an issue when opening both the master and slave ends from the
same process on Node though. When doing so Node will hang at exit
because of stuck read calls in the eventloop. I wasn't able to figure
out the cause, so a hack I used was to open the slave end into a
subprocess that we kill once done with it. This only impacts our test
suite.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>