Make tests resilent across gcc/gdb versions
Different combinations of gdb/gcc consider a function breakpoint
on different lines.

Older versions stop on the first line of code, newer versions
stop on the opening brace.

This change makes the opening brace and first statement the same
line of code so tests can be agnostic to this difference.