Issue #178: Add more escaping on Windows for old GDB
The handling of \ on Windows causes lots of minor issues. While
CDT only does the double-escaping of paths if there is whitespace,
on cdt-gdb-adapter we need to do it all the time as the effect
on original-location is different. In CDT we don't worry about
original-location, but in cdt-gdb-adapter we need to be able to
identify which DAP breakpoint location matches which GDB breakpoint
reliably.