Use bultin chai match so that test failures diplay well
This change avoids losing information from the test. Without this
change if a test failed the error would be:

     The display value of the array is not a hexadecimal address
      + expected - actual

With this change it is the more useful:

     AssertionError: The display value of the array is not a hexadecimal
     address: expected '0xc38cdffa14' to match /0x[\d]+/

With such an error message it would have been instantaneous to
identify that the issue was a bad regex and not something more
serious.