Move common launch arguments to utils
Up until this change we have had to update every test individually
to add a new option to the test matrix. This change adds a new
way to create the various launch arguments so that only
one place needs to be updated in the future.

The test argument to fillDefaults is to simplify some future
improvements to the logging coming in subsequent commits.