Allow passing command line arguments to adapter without shell
Remove shell=true when spawning adapter in tests

The shell=true was a workaround so that arguments could be
passed to node. But it leads to platform specific handling.
Therefore remove the shell=true which also means the workarounds
that would have been needed to quote JSON on the command line
properly are not needed.

Needed to test #227