Increase timeout on Windows to 25s in previously missed locations
Most of the timeouts for tests are the mocha timeouts, but DebugClient
has its own timeout values. This change synchronizes those values.

This increased timeout is specifically needed for the attachRemote
test which takes more than 5 seconds to do the initial connection
on the GitHub actions build regularly.

This is a follow up to commit 3708cea2e0833e0d48fbcc7ca1e4718894c9809e
part of #222