Rework waitForOutputEvent to not miss events
The code used to do this.once('output') in a tight loop,
but this meant that between processing one of the iterations
and the next iteration of the loop a message may have
been output. Instead use this.on('output') so all messages
are processed.

Prerequisite for #239 which generates multiple events quickly