Wait GDB server is terminated before start new one
Currently, the launch method does not wait for the Disconnect method to finish to start a new GDB server. The new GDB server could not be started because the old one was still holding the emulator when the debug session restart was requested. The solution is to wait until the GDB server exits then launch will be called.