Update status of multi threads in Call Stack View (#257)
For multiple cores debugging, the state of cores is not updated correctly with the state in the GDB server.
Cannot use "Resume" or "Suspend" button in case of multiple cores debugging. That has not happened in
cdt-amalgamator. This solution refers to cdt-amalgamator repository.

Root cause: Call Stack View does not update the state RUNNING if the process *Initialize* is not completed.
cdt-gdb-adapter does not support sending "ContinuedEvent" when GDB returns async message "running"

Solution: Return Continues Response with the arg "allThreadsContinued" is 'false', which makes all threads
not set to "RUNNING" state. Send ContinuedEvent when handling async message "running" from the GDB server.
Remove the condition that does not send StoppedEvent if the process Initialize is not completed.