Allow attaching (remote and local) with no program
There is no strict need to have a program passed to GDB when
attaching (remote or local). Either the program can be downloaded
from the remote or otherwise read by GDB, or if not possible
then GDB can do symbol-free debugging.

Previously in #228 some work was done to better error when
program was not specified. This change loosens the restrictions
slightly. If the program is not specified, it is still a warning
in VSCode because of
[package.json](https://github.com/eclipse-cdt-cloud/cdt-gdb-vscode/blob/22a9f1048f9030689ed1f5f2b3e5d5e48df9047d/package.json#L205-L207)

See this comment for how it is presented to a user:
https://github.com/eclipse-cdt-cloud/cdt-gdb-adapter/issues/261#issuecomment-1505209833

Part of #261