Support Debug Console interacting directly with GDB
This change adds the ability to interact with GDB
directly in the debug console of VSCode by providing
a way to bypass the normal expression evaluation
and sending commands to the GDB console input.

To display the value of an expression, type that
expression which can reference variables that are in
scope. For example type '2 + 3' or the name of a
variable.
Arbitrary commands can be sent to GDB by prefixing
the input with a '>', for example type '>show version'
or '>help'.