Disconnect before exit for remote targets (#273)
This changes fixes an issue whereby when debugging on a remote target,
the exit sequence is not entirely correct. We should send a "disconnect"
GDB command before "exit". This prevents a messages such as "Remote
doesn't know how to detach" and "Cannot execute this command while the
target is running" from appearing, as they can be somewhat misleading.