Improvements for Breakpoint Logic (#268)
In this update there are update for breakpoint insert logic.

We introduce a new method getBreakpointOptions to GDBBackend class.
With this new functionality, any child class could override the
breakpoint options (including changing the type hardware/software).

This will enable a new feature for embedded systems to manage the
limited breakpoint capability more efficiently. Also, any error thrown
in the getBreakpointOptions will disable the breakpoint, thus, limited
capabilities controlled without having errors in the user interface.

There are also some refactoring performed during the operation. Perhaps,
not all the refactorings are must to have, but tried to handle the
namings of functions/methods and object structure to be semantically
coherent after the update.