Adding the ability to configure serial port output to the debug console (#271)
This commit introduces a set of parameters that allow us to configure a
serial line in the adapter to print UART output from on-board/emulation
debug to the debug console. These parameters include the TCP port to
hook onto (for emulation programs outputting UART output), or the serial
port on the host-machine to capture UART output from. For the serial
line, we also configure the baud rate, stop bits, handshaking method,
etc.

The source code looks to take this parameters to then configure serial
ports and sockets to read UART data and output it to the debug console.
It will also set up necessary event handlers, and changes the log level
in the non-verbose case to allow for "log" messages to be printed in a
manner such that UART output isn't designated as a "warning".