Add threadId and FrameId as optional parameters
The future needs to read/write memory on a thread basis requires
the GDB MI commands to be updated.

The GDB MI commands can take a threadId and frameId if specified
so add that capability as an optional parameter.

Add unit test to verify threadId and frameId changes. Multithread
test is needed to showcase this change.