Remove unnecessary check in test
The UART fails to open test was looking for a `0` in the error
message, corresponding to the port number that was being opened.
However, depending on the specific machine run on, the actual
error from the backend may vary. On GitHub actions instead of
referring to the port number it says:

  "error on socket connection\n - AggregateError"

And on my local machine it shows

  "error on socket connection\n - Error: connect ECONNREFUSED 127.0.0.1"

With this change we still require the adapter to report an error
on socket connection, but the specific error details are not
required to match any particular pattern.