Disassembly address handling improvement (#348)
We observe some issues in while using the disassembly in embedded devices due to sending disassemble request for negative memory regions. We implemented a code-update to limit the lower bound of the query to the address 0x0.

If we still need to fill more instructions depending on the DAP request, you may realise that we used relative addresses in the empty instructions (e.g. (0x00)-2, (0x00)-4 etc). This is arguable and I am happy to hear your thoughts. Simply the VSCode is ignoring these lines and not showing in the disassembly window (maybe because they are not convertable to number or bigint).

It was hard to check this behaviour in unit test, thus I implemented a test for the internal function and checked the behaviour over that internal function.