Hardware/Software Breakpoint Modes
In this implementation, we introduced the "Breakpoint Mode" usage,
which is recently introduced to the DAP. This implementation sends
the breakpoint mode capabilities at the initialisation phase of the
debug adapter. We implemented the following modes:

* hardware: The mode `hardware` indicates the hardware breakpoints.
* software: The mode `software` indicates the software breakpoints.

In VSCode, user can choose the breakpoint mode after adding the
breakpoint by editing the breakpoint details.

We also introduce `mode` property to the `MIBreakpointInsertOptions`
interface. If there is a value in the `mode` property, the breakpoint
insert operation is prioritising the value of the `mode` over the
`hardware` property. Additionally, `hardware` property flagged as
deprecate and planning to be removed soon.