Fix getting stuck on concurrent breakpoint setup (1/3)
When several calls to pauseIfNeeded arrive in quick succession from
different set*BreakpointRequests during the configuration phase, the
debug session would hang and never complete some of those requests. This
is because when the second call to pauseIfNeeded arrives, it overwrites
the previous waitPaused resolve function, and the first call then waits
indefinitely on a promise that never resolves because its resolve
function was lost.

Fix this by daisy-chaining the previous resolve function on the new
promise.

Signed-off-by: Christian Walther <walther@indel.ch>