Fix extra continue request and redundant unpausing and re-pausing (3/3)
When attaching to a target that is not stopped by attaching, there is a
superfluous sendExecContinue at the end of the configuration phase,
whose error response from GDB shuts down the debug session. The target
is already running at this point and does not need another "continue",
thanks to the balanced pauseIfNeeded/continueIfNeeded calls ensured by
the reference count from the previous commit.

Additionally, if set*BreakpointRequests were to arrive not concurrently
but sequentially (which they do not in my experience, but the protocol
would allow it), with all previous ones complete before the next one
arrives, we would still unnecessarily pause and continue multiple times,
because the refcount would go to zero in between.

Both issues are fixed as follows: During the configuration phase, if and
when the first request that requires pausing the target arrives, add
another outer bracket of pauseIfNeeded/continueIfNeeded around the whole
remainder of the phase, so that the refcount never goes to zero in
between. Subsume the previously extraneous "continue" into that final
continueIfNeeded. When the target is initially paused, start at refcount
1 and the corresponding state, that way things continue to work with
targets that do stop on attaching.

The start of the configuration phase (sending of the "initialized"
event) was previously handled separately in some debug session
subclasses. Since we are now adding additional processing at that point,
centralize it by introducing a superclass method "sendInitializedEvent".

This finally makes the test introduced 3 commits back pass.

Signed-off-by: Christian Walther <walther@indel.ch>