Add test for concurrent breakpoint setup, currently failing
The current pauseIfNeeded implementation is incapable of handling
concurrent breakpoint setting requests, as sent by Visual Studio Code.
The problem only manifests with a target that does not automatically
stop all threads when attaching - unlike what gdb and gdbserver do when
attaching to a Unix process, but often required with embedded systems,
where stopping would disrupt real-time behavior. Simulate such a target
by issuing an -exec-continue right after attaching.

The test program for the "attach remote" suite is changed from "empty"
to "loopforever", which does not disturb the existing tests, but is
required for the new one, which needs a target that does not immediately
hit a breakpoint after the initial "continue", but stays running until
explicitly steered onto the breakpoint.

The new test fails on Windows when gdb-async is off: The first attempt
to interrupt the target results in

  GDB signal: SIGINT to pid 1424
  To client: {"seq":0,"type":"event","event":"output","body":{"category":"server","output":"gdb killed by signal SIGINT\n"}}
  GDB Remote session: GDB exited with code null, signal SIGINT

I have not found any information on whether that is a fundamental
limitation of GDB on Windows, but all the other tests that attempt the
same thing are skipped under those circumstances, so do the same.

Signed-off-by: Christian Walther <walther@indel.ch>