Add test for detaching after running program exited, currently failing
When the debugged program has exited after it was, to the adapter's
knowledge, running, a `disconnect` request gets stuck and never
completes, because `GDBTargetDebugSession.doDisconnectRequest` tries to
pause the program before issuing the `disconnect` command to GDB, but
the stop notification that it expects as confirmation that the pause has
taken effect never arrives, because there are no threads anymore that
could stop.

This happens here in the afterEach function which first kills the
gdbserver, causing GDB to consider the program exited, and then in
`DebugClient.stop()` sends a disconnect request. It causes the test to
fail with a timeout in the "after each" hook.

The problem was not noticed in the existing tests because they all
happen to end with the program suspended, so no pausing is attempted.

Signed-off-by: Christian Walther <walther@indel.ch>