Fix detach getting stuck on exited program
When attempting to pause a target that has no threads anymore because
the program has exited in the meantime, don't wait for a stop
notification that will never come.

There is still a race condition in that the threads could disappear
after we ask for them, but the likelihood is reduced.

This makes the test from the previous commit pass.

Signed-off-by: Christian Walther <walther@indel.ch>