Improvement for step functions latencies
The step functions were providing poor UX in slow debugging sessions.
The debug buttons were unresponsive for longer times cause users to
repeatedly click on stepping buttons which leads to multiple unintended
step operations.

In this implementation, a timeouted response is returned to the user
interface to change the UI state properly. For this purpose, we
implemented:

- A new utility function called `sendResponseWithTimeout` to handle the
  operation.
- Improved the `nextRequest`, `stepInRequest` and `stepOutRequest`
  methods to introduce the timeouted response, with a default of 100ms
  timeout.
- Introduce a new launch variable called `steppingResponseTimeout` to
  provide an option that user can change the default timeout for
  stepping functions.