Fix getting stuck on concurrent breakpoint setup (#407)
* Add test for concurrent breakpoint setup, currently failing

The current pauseIfNeeded implementation is incapable of handling
concurrent breakpoint setting requests, as sent by Visual Studio Code.
The problem only manifests with a target that does not automatically
stop all threads when attaching - unlike what gdb and gdbserver do when
attaching to a Unix process, but often required with embedded systems,
where stopping would disrupt real-time behavior. Simulate such a target
by issuing an -exec-continue right after attaching.

The test program for the "attach remote" suite is changed from "empty"
to "loopforever", which does not disturb the existing tests, but is
required for the new one, which needs a target that does not immediately
hit a breakpoint after the initial "continue", but stays running until
explicitly steered onto the breakpoint.

The new test fails on Windows when gdb-async is off: The first attempt
to interrupt the target results in

  GDB signal: SIGINT to pid 1424
  To client: {"seq":0,"type":"event","event":"output","body":{"category":"server","output":"gdb killed by signal SIGINT\n"}}
  GDB Remote session: GDB exited with code null, signal SIGINT

I have not found any information on whether that is a fundamental
limitation of GDB on Windows, but all the other tests that attempt the
same thing are skipped under those circumstances, so do the same.

Signed-off-by: Christian Walther <walther@indel.ch>

* Fix getting stuck on concurrent breakpoint setup (1/3)

When several calls to pauseIfNeeded arrive in quick succession from
different set*BreakpointRequests during the configuration phase, the
debug session would hang and never complete some of those requests. This
is because when the second call to pauseIfNeeded arrives, it overwrites
the previous waitPaused resolve function, and the first call then waits
indefinitely on a promise that never resolves because its resolve
function was lost.

Fix this by daisy-chaining the previous resolve function on the new
promise.

Signed-off-by: Christian Walther <walther@indel.ch>

* Fix redundant pausing and early unpausing during concurrent breakpoint setup (2/3)

When several set*BreakpointRequests arrive in quick succession, their
pauseIfNeeded calls redundantly issue the same number of thread info
requests and pause requests to GDB in a row when one would be
sufficient, and the first continueIfNeeded call from whichever of the
breakpoint requests completes first already ends the pause even though
the rest are not done with their work yet.

Fix this by introducing a reference count of active pause requests, so
that only the first call to pauseIfNeeded will start a pause, and only
the last call to continueIfNeeded will end it.

Signed-off-by: Christian Walther <walther@indel.ch>

* Fix extra continue request and redundant unpausing and re-pausing (3/3)

When attaching to a target that is not stopped by attaching, there is a
superfluous sendExecContinue at the end of the configuration phase,
whose error response from GDB shuts down the debug session. The target
is already running at this point and does not need another "continue",
thanks to the balanced pauseIfNeeded/continueIfNeeded calls ensured by
the reference count from the previous commit.

Additionally, if set*BreakpointRequests were to arrive not concurrently
but sequentially (which they do not in my experience, but the protocol
would allow it), with all previous ones complete before the next one
arrives, we would still unnecessarily pause and continue multiple times,
because the refcount would go to zero in between.

Both issues are fixed as follows: During the configuration phase, if and
when the first request that requires pausing the target arrives, add
another outer bracket of pauseIfNeeded/continueIfNeeded around the whole
remainder of the phase, so that the refcount never goes to zero in
between. Subsume the previously extraneous "continue" into that final
continueIfNeeded. When the target is initially paused, start at refcount
1 and the corresponding state, that way things continue to work with
targets that do stop on attaching.

The start of the configuration phase (sending of the "initialized"
event) was previously handled separately in some debug session
subclasses. Since we are now adding additional processing at that point,
centralize it by introducing a superclass method "sendInitializedEvent".

This finally makes the test introduced 3 commits back pass.

Signed-off-by: Christian Walther <walther@indel.ch>

* Whitespace rewrapping in increased indentation by prettier

Signed-off-by: Christian Walther <walther@indel.ch>

* Changelog

Signed-off-by: Christian Walther <walther@indel.ch>

---------

Signed-off-by: Christian Walther <walther@indel.ch>